requirements.txt
Flask==2.0.1
tweepy==4.10.0
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
serverless-wsgi==3.0.2
google-api-python-client==2.97.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
nltk==3.8.1
